Refactor renderer cvar registration by subsystem
