+++ Quake/gl_rmisc.c
@@ -27,6 +27,8 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "mat_shader.h"
 #include "r_dlight_pool.h"
 #include "r_postfx.h"
+#include "r_ssao.h"
+#include "r_godrays.h"
 #include "r_fogvol.h"
 
 //johnfitz -- new cvars
@@ -121,124 +123,6 @@ extern void TexMgr_SRGBTextures_f (cvar_t *var);
 extern void TexMgr_LightmapColorspace_f (cvar_t *var);
 extern void TexMgr_LightmapLinearCompat_f (cvar_t *var);
 extern cvar_t r_bloom_threshold;
-extern cvar_t r_postfx;
-extern cvar_t r_polyblend_legacy;
-extern cvar_t r_postfx_pickup;
-extern cvar_t r_postfx_pickup_exposure;
-extern cvar_t r_postfx_pickup_bloom;
-extern cvar_t r_postfx_pickup_duration;
-extern cvar_t r_postfx_damage;
-extern cvar_t r_postfx_damage_vignette;
-extern cvar_t r_postfx_damage_vignette_softness;
-extern cvar_t r_postfx_damage_desat;
-extern cvar_t r_postfx_damage_exposure;
-extern cvar_t r_postfx_damage_duration;
-extern cvar_t r_postfx_damage_accum_window;
-extern cvar_t r_postfx_damage_accum_scale;
-extern cvar_t r_post_damage_doublevision;
-extern cvar_t r_post_damage_dv_strength;
-extern cvar_t r_post_damage_dv_px;
-extern cvar_t r_post_damage_dv_freq;
-extern cvar_t r_post_damage_trauma_scale;
-extern cvar_t r_post_damage_trauma_decay;
-extern cvar_t r_post_damage_dv_quality;
-extern cvar_t r_post_damage_dv_debug;
-extern cvar_t r_postfx_powerup;
-extern cvar_t r_postfx_powerup_lut_strength;
-extern cvar_t r_postfx_powerup_ramp_in;
-extern cvar_t r_postfx_powerup_ramp_out;
-extern cvar_t r_postfx_underwater;
-extern cvar_t r_postfx_underwater_grade_strength;
-extern cvar_t r_postfx_underwater_fog_strength;
-extern cvar_t r_postfx_underwater_ramp_in;
-extern cvar_t r_postfx_underwater_ramp_out;
-extern cvar_t r_postfx_underwater_fog_water_r;
-extern cvar_t r_postfx_underwater_fog_water_g;
-extern cvar_t r_postfx_underwater_fog_water_b;
-extern cvar_t r_postfx_underwater_fog_slime_r;
-extern cvar_t r_postfx_underwater_fog_slime_g;
-extern cvar_t r_postfx_underwater_fog_slime_b;
-extern cvar_t r_postfx_underwater_fog_lava_r;
-extern cvar_t r_postfx_underwater_fog_lava_g;
-extern cvar_t r_postfx_underwater_fog_lava_b;
-extern cvar_t r_postfx_quad;
-extern cvar_t r_postfx_quad_emissive_boost;
-extern cvar_t r_postfx_quad_bloom_boost;
-extern cvar_t r_postfx_quad_pulse_speed;
-extern cvar_t r_postfx_quad_pulse_intensity;
-extern cvar_t r_postfx_bloom_mode;
-extern cvar_t r_postfx_lut;
-extern cvar_t r_postfx_lut_strength_powerup;
-extern cvar_t r_postfx_lut_strength_underwater;
-extern cvar_t r_postfx_lut_debug_id;
-extern cvar_t r_postfx_debug;
-extern cvar_t r_ssao;
-extern cvar_t r_ssao_radius;
-extern cvar_t r_ssao_intensity;
-extern cvar_t r_ssao_bias;
-extern cvar_t r_ssao_power;
-extern cvar_t r_ssao_min;
-extern cvar_t r_ssao_samples;
-extern cvar_t r_ssao_blur;
-extern cvar_t r_ssao_blur_radius;
-extern cvar_t r_ssao_blur_sigma;
-extern cvar_t r_ssao_blur_bilateral;
-extern cvar_t r_ssao_halfres;
-extern cvar_t r_ssao_debug;
-extern cvar_t r_ssao_debug_far;
-extern cvar_t r_ssao_reversedz_mode;
-extern cvar_t r_ssao_noise;
-extern cvar_t r_ssao_noise_mode;
-extern cvar_t r_ssao_noise_scale;
-extern cvar_t r_ssao_normalsource;
-extern cvar_t r_ssao_freeze_noise;
-extern cvar_t r_ssao_force_fullres;
-extern cvar_t r_ssao_format;
-extern cvar_t r_ssao_upscale_nearest;
-extern cvar_t r_ssao_fog_strength;
-extern cvar_t r_ssao_fog_power;
-extern cvar_t r_ssao_max_distance;
-extern cvar_t r_ssao_validate;
-extern cvar_t r_godrays;
-extern cvar_t r_godrays_emit_emissive;
-extern cvar_t r_godrays_emit_lighttex;
-extern cvar_t r_godray_sky_enable;
-extern cvar_t r_godray_sky_threshold;
-extern cvar_t r_godray_sky_intensity;
-extern cvar_t r_godray_sky_blur;
-extern cvar_t r_godrays_sky_enable;
-extern cvar_t r_godrays_sky_threshold;
-extern cvar_t r_godrays_sky_intensity;
-extern cvar_t r_godrays_sky_tint;
-extern cvar_t r_godrays_emissive_intensity;
-extern cvar_t r_godrays_lighttex_intensity;
-extern cvar_t r_godrays_emissive_threshold;
-extern cvar_t r_godrays_light_threshold;
-extern cvar_t r_godrays_mask_knee;
-extern cvar_t r_godrays_blur;
-extern cvar_t r_godrays_lighttex_name_match;
-extern cvar_t r_godrays_samples;
-extern cvar_t r_godrays_density;
-extern cvar_t r_godrays_weight;
-extern cvar_t r_godrays_decay;
-extern cvar_t r_godrays_exposure;
-extern cvar_t r_godrays_threshold;
-extern cvar_t r_godrays_sky_softness;
-extern cvar_t r_godrays_light_sharpness;
-extern cvar_t r_godrays_max_radius;
-extern cvar_t r_godrays_light_x;
-extern cvar_t r_godrays_light_y;
-extern cvar_t r_godrays_stabilize;
-extern cvar_t r_godrays_stabilize_strength;
-extern cvar_t r_godrays_stabilize_max_px;
-extern cvar_t r_godrays_smooth_rate;
-extern cvar_t r_godrays_max_shift;
-extern cvar_t r_godrays_reset_on_teleport;
-extern cvar_t r_godrays_debug;
-extern cvar_t r_godrays_debug_source;
-extern cvar_t r_godrays_volumetric;
-extern cvar_t r_godrays_vol_pow;
-extern cvar_t r_sun_light;
 extern cvar_t r_vignette;
 extern cvar_t r_vignette_radius_inner;
 extern cvar_t r_vignette_radius_outer;
@@ -250,6 +134,7 @@ extern cvar_t r_vignette_blend_mode;
 extern cvar_t r_vignette_noise;
 extern cvar_t r_teleportfx;
 extern cvar_t r_teleportfx_time;
+extern cvar_t r_sun_light;
 
 #if defined(USE_SIMD)
 extern cvar_t r_simd;
@@ -520,16 +405,6 @@ float GL_WaterAlphaForTextureType (textype_t type)
 }
 
 
-static void R_GodraysMigrateLegacySkyToCanonical (cvar_t *unused)
-{
-	(void)unused;
-
-	Cvar_SetQuick (&r_godrays_sky_enable, r_godray_sky_enable.string);
-	Cvar_SetQuick (&r_godrays_sky_threshold, r_godray_sky_threshold.string);
-	Cvar_SetQuick (&r_godrays_sky_intensity, r_godray_sky_intensity.string);
-	Cvar_SetQuick (&r_godrays_sky_softness, r_godray_sky_blur.string);
-}
-
 
 /*
 ===============
@@ -639,131 +514,10 @@ Cvar_RegisterVariable (&r_color_contrast);
 Cvar_RegisterVariable (&r_color_saturation);
 Cvar_RegisterVariable (&r_bloom);
 Cvar_RegisterVariable (&r_bloom_threshold);
-Cvar_RegisterVariable (&r_postfx);
-Cvar_RegisterVariable (&r_polyblend_legacy);
-Cvar_RegisterVariable (&r_postfx_pickup);
-Cvar_RegisterVariable (&r_postfx_pickup_exposure);
-Cvar_RegisterVariable (&r_postfx_pickup_bloom);
-Cvar_RegisterVariable (&r_postfx_pickup_duration);
-Cvar_RegisterVariable (&r_postfx_damage);
-Cvar_RegisterVariable (&r_postfx_damage_vignette);
-Cvar_RegisterVariable (&r_postfx_damage_vignette_softness);
-Cvar_RegisterVariable (&r_postfx_damage_desat);
-Cvar_RegisterVariable (&r_postfx_damage_exposure);
-Cvar_RegisterVariable (&r_postfx_damage_duration);
-Cvar_RegisterVariable (&r_postfx_damage_accum_window);
-Cvar_RegisterVariable (&r_postfx_damage_accum_scale);
-Cvar_RegisterVariable (&r_post_damage_doublevision);
-Cvar_RegisterVariable (&r_post_damage_dv_strength);
-Cvar_RegisterVariable (&r_post_damage_dv_px);
-Cvar_RegisterVariable (&r_post_damage_dv_freq);
-Cvar_RegisterVariable (&r_post_damage_trauma_scale);
-Cvar_RegisterVariable (&r_post_damage_trauma_decay);
-Cvar_RegisterVariable (&r_post_damage_dv_quality);
-Cvar_RegisterVariable (&r_post_damage_dv_debug);
-Cvar_RegisterVariable (&r_postfx_powerup);
-Cvar_RegisterVariable (&r_postfx_powerup_lut_strength);
-Cvar_RegisterVariable (&r_postfx_powerup_ramp_in);
-Cvar_RegisterVariable (&r_postfx_powerup_ramp_out);
-Cvar_RegisterVariable (&r_postfx_underwater);
-Cvar_RegisterVariable (&r_postfx_underwater_grade_strength);
-Cvar_RegisterVariable (&r_postfx_underwater_fog_strength);
-Cvar_RegisterVariable (&r_postfx_underwater_ramp_in);
-Cvar_RegisterVariable (&r_postfx_underwater_ramp_out);
-Cvar_RegisterVariable (&r_postfx_underwater_fog_water_r);
-Cvar_RegisterVariable (&r_postfx_underwater_fog_water_g);
-Cvar_RegisterVariable (&r_postfx_underwater_fog_water_b);
-Cvar_RegisterVariable (&r_postfx_underwater_fog_slime_r);
-Cvar_RegisterVariable (&r_postfx_underwater_fog_slime_g);
-Cvar_RegisterVariable (&r_postfx_underwater_fog_slime_b);
-Cvar_RegisterVariable (&r_postfx_underwater_fog_lava_r);
-Cvar_RegisterVariable (&r_postfx_underwater_fog_lava_g);
-Cvar_RegisterVariable (&r_postfx_underwater_fog_lava_b);
-Cvar_RegisterVariable (&r_postfx_quad);
-Cvar_RegisterVariable (&r_postfx_quad_emissive_boost);
-Cvar_RegisterVariable (&r_postfx_quad_bloom_boost);
-Cvar_RegisterVariable (&r_postfx_quad_pulse_speed);
-Cvar_RegisterVariable (&r_postfx_quad_pulse_intensity);
-Cvar_RegisterVariable (&r_postfx_bloom_mode);
-Cvar_RegisterVariable (&r_postfx_lut);
-Cvar_SetCallback (&r_postfx_lut, R_PostFX_ReloadLUTs_f);
-Cvar_RegisterVariable (&r_postfx_lut_strength_powerup);
-Cvar_RegisterVariable (&r_postfx_lut_strength_underwater);
-Cvar_RegisterVariable (&r_postfx_lut_debug_id);
-Cvar_RegisterVariable (&r_postfx_debug);
-Cvar_RegisterVariable (&r_ssao);
-Cvar_RegisterVariable (&r_ssao_radius);
-Cvar_RegisterVariable (&r_ssao_intensity);
-Cvar_RegisterVariable (&r_ssao_bias);
-Cvar_RegisterVariable (&r_ssao_power);
-Cvar_RegisterVariable (&r_ssao_min);
-Cvar_RegisterVariable (&r_ssao_samples);
-	Cvar_RegisterVariable (&r_ssao_blur);
-	Cvar_RegisterVariable (&r_ssao_blur_radius);
-	Cvar_RegisterVariable (&r_ssao_blur_sigma);
-	Cvar_RegisterVariable (&r_ssao_blur_bilateral);
-	Cvar_RegisterVariable (&r_ssao_halfres);
-	Cvar_RegisterVariable (&r_ssao_debug);
-	Cvar_RegisterVariable (&r_ssao_debug_far);
-	Cvar_RegisterVariable (&r_ssao_reversedz_mode);
-	Cvar_RegisterVariable (&r_ssao_noise);
-	Cvar_RegisterVariable (&r_ssao_noise_mode);
-	Cvar_RegisterVariable (&r_ssao_noise_scale);
-	Cvar_RegisterVariable (&r_ssao_normalsource);
-	Cvar_RegisterVariable (&r_ssao_freeze_noise);
-	Cvar_RegisterVariable (&r_ssao_force_fullres);
-	Cvar_RegisterVariable (&r_ssao_format);
-	Cvar_RegisterVariable (&r_ssao_upscale_nearest);
-	Cvar_RegisterVariable (&r_ssao_fog_strength);
-	Cvar_RegisterVariable (&r_ssao_fog_power);
-	Cvar_RegisterVariable (&r_ssao_max_distance);
-	Cvar_RegisterVariable (&r_ssao_validate);
-Cvar_RegisterVariable (&r_godrays);
-	Cvar_RegisterVariable (&r_godrays_emit_emissive);
-	Cvar_RegisterVariable (&r_godrays_emit_lighttex);
-	Cvar_RegisterVariable (&r_godray_sky_enable);
-	Cvar_RegisterVariable (&r_godray_sky_threshold);
-	Cvar_RegisterVariable (&r_godray_sky_intensity);
-	Cvar_RegisterVariable (&r_godray_sky_blur);
-	Cvar_RegisterVariable (&r_godrays_sky_enable);
-	Cvar_RegisterVariable (&r_godrays_sky_threshold);
-	Cvar_RegisterVariable (&r_godrays_sky_intensity);
-	Cvar_RegisterVariable (&r_godrays_sky_tint);
-	Cvar_RegisterVariable (&r_godrays_emissive_intensity);
-	Cvar_RegisterVariable (&r_godrays_lighttex_intensity);
-	Cvar_RegisterVariable (&r_godrays_emissive_threshold);
-	Cvar_RegisterVariable (&r_godrays_light_threshold);
-	Cvar_RegisterVariable (&r_godrays_mask_knee);
-	Cvar_RegisterVariable (&r_godrays_blur);
-Cvar_RegisterVariable (&r_godrays_lighttex_name_match);
-Cvar_RegisterVariable (&r_godrays_samples);
-Cvar_RegisterVariable (&r_godrays_density);
-Cvar_RegisterVariable (&r_godrays_weight);
-Cvar_RegisterVariable (&r_godrays_decay);
-Cvar_RegisterVariable (&r_godrays_exposure);
-Cvar_RegisterVariable (&r_godrays_threshold);
-Cvar_RegisterVariable (&r_godrays_sky_softness);
-	/* Legacy r_godray_sky_* aliases are parse/set shims that write canonical r_godrays_sky_* values. */
-	Cvar_SetCallback (&r_godray_sky_enable, R_GodraysMigrateLegacySkyToCanonical);
-	Cvar_SetCallback (&r_godray_sky_threshold, R_GodraysMigrateLegacySkyToCanonical);
-	Cvar_SetCallback (&r_godray_sky_intensity, R_GodraysMigrateLegacySkyToCanonical);
-	Cvar_SetCallback (&r_godray_sky_blur, R_GodraysMigrateLegacySkyToCanonical);
-	/* Startup migration: values restored into legacy names seed canonical cvars once. */
-	R_GodraysMigrateLegacySkyToCanonical (NULL);
-Cvar_RegisterVariable (&r_godrays_light_sharpness);
-Cvar_RegisterVariable (&r_godrays_max_radius);
-Cvar_RegisterVariable (&r_godrays_light_x);
-Cvar_RegisterVariable (&r_godrays_light_y);
-Cvar_RegisterVariable (&r_godrays_stabilize);
-Cvar_RegisterVariable (&r_godrays_stabilize_strength);
-Cvar_RegisterVariable (&r_godrays_stabilize_max_px);
-Cvar_RegisterVariable (&r_godrays_smooth_rate);
-	Cvar_RegisterVariable (&r_godrays_max_shift);
-	Cvar_RegisterVariable (&r_godrays_reset_on_teleport);
-	Cvar_RegisterVariable (&r_godrays_debug);
-	Cvar_RegisterVariable (&r_godrays_debug_source);
-	Cvar_RegisterVariable (&r_godrays_volumetric);
-	Cvar_RegisterVariable (&r_godrays_vol_pow);
+	/* CVar owners: r_postfx* in r_postfx.c, r_ssao* in r_ssao.c, r_godray* and r_godrays* in r_godrays.c. */
+	R_PostFX_RegisterCvars ();
+	R_SSAO_RegisterCvars ();
+	R_Godrays_RegisterCvars ();
 Cvar_RegisterVariable (&r_vignette);
 	Cvar_RegisterVariable (&r_vignette_radius_inner);
 	Cvar_RegisterVariable (&r_vignette_radius_outer);
@@ -827,6 +581,8 @@ Cvar_RegisterVariable (&r_vignette);
 	Cvar_SetCallback (&r_telealpha, R_SetTelealpha_f);
 	Cvar_SetCallback (&r_slimealpha, R_SetSlimealpha_f);
 
+	/* CVar owner: r_fogvol* in r_fogvol.c. */
+	R_FogVol_RegisterCvars ();
 	R_PostFX_Init ();
 	R_FogVol_Init ();
