+++ Quake/r_ssao.h
@@ -11,5 +11,6 @@ typedef struct r_ssao_fog_state_s
 
 float R_SSAO_SanitizeValue (float value, float fallback, float minval, float maxval);
 void R_SSAO_CaptureFogState (const gpuframedata_t *framedata, r_ssao_fog_state_t *out_state);
+void R_SSAO_RegisterCvars (void);
 
 #endif
